Update Player class, create utils.py
Player can now be created by only passing a player's ID instead
of having to pass an additional dictionary. Also moved some
functions to utils.py, as they are needed in other classes as
well.
